Updated SnakeBossState Address
Possible fix for final boss split, with an updated address for SnakeBossState, which has worked on multiple installs and PC's, and has worked from start to end in a single full run. Also has some other small code changes related to the split its-self.

Either it works, or it does nothing different than it already was, so worth pushing.