Timer Resume at Menu, Added Splits, Fixes, and general clean up.
Adds variable "splashOver" which sets to false when the game closes, and is only set to true when it first gets a loading screen, during which time it will return "isLoading" to True, blocking timer from resuming for the time until it reaches menu.

Adds Split for Chapter 1 Mataan Entrance and Chapter 3 Mall Entrance. 

Fixes Splits for Pyramid Entrance and Versum Hill Entrance.

Modified the autosplit menu layout to be more readable.

Moved code around to be more readable.